Add -device-code CLI option to use old oauth flow
Make the PKCE OAuth2 flow the default but allow users to use the
previous device-code flow via a CLI flag. Preserve the old print or open
behavior based on the -stdout flag.

Await the callback indefinitely in the PKCE flow vs the previous 1
minute timeout. Signal errors via channel back to the caller to enable
quicker exiting if there is a callback exception.