Add AWS_VAULT_OAUTH_CALLBACK_PORT env var & CLI flag
Allow users to specify the port that the OAuth2 callback server listens
on rather than forcing it to be chosen at random. This allows for
deployments in remote workstations using aws-vault in combination with
SSH tunnels created on known ports between the client and remote
machines to perform browser-based logins.